Infra.Automation-monitoring/requirements.txt
urllib3==1.25.9
requests==2.26.0
PyYAML==5.3.1
jenkinsapi==0.3.11
python-jenkins==1.0.1

tabulate
beautifulsoup4